requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma
chromadb
pydantic
python-dotenv
streamlit
